Change response to standard response

+++ src/main/java/com/iduy/gondrong/repository/ProposalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProposalRepository extends JpaRepository<Proposal,Integer> {
+public interface ProposalRepository extends JpaRepository<Proposal,Long> {
 }
